GH-42 Refactor: move teststrategy base class
Having the abstract base class in the same namespace
as its concrete implementations makes instantiating
the test classes more complicated.

I therefore moved the base class one level up.
